requirements.txt
flask==3.0.2
tensorflow==2.15.0
numpy==1.24.3
opencv-python==4.9.0.80
pillow==10.2.0
flask-cors==4.0.0
werkzeug==3.0.1 